#48 Feature : 티켓 생성 구현 완료 - Cache 에서 티켓 확인 후 break 추가 필요 - 캐시 value 가 null 로 저장되는 현상 수정 필요
- Category 에서 orderBy 제거하니 성공함.
- Category 의 findByPageable 확인 필요.
