+++ src/main/kotlin/com/codersgate/ticketraider/domain/ticket/dto/SeatInfo.kt
@@ -1,0 +1,8 @@
+package com.codersgate.ticketraider.domain.ticket.dto
+
+import com.codersgate.ticketraider.domain.ticket.entity.TicketGrade
+
+data class SeatInfo(
+    val ticketGrade: TicketGrade,
+    val seatNumber: Int
+)
